Define fallback version
The `--version` option relies on `git describe` command to determine the
software version. This requires the script to be located within a valid
git repository, which may not always be the case.

This patch tries to determine the software version using git and adds a
fallback value (`unknown`) in case this is not possible.

Using the variable `_VERSION_PLACEHOLDER` allows third party tools such
as Ansible to edit this specific line with the proper version at the
time of deployment.