Getting WebDAV door from the API (#98)
* Getting WebDAV door from the API
This commit adds a function `get_webdav_door` that allows Ada to get the URL of a WebDAV door from the API.

Ada does not know about WebDAV doors; it only knows the API's URL.

There are two ways you can ask the API for a WebDAV door. Both methods are implemented.
If the first method fails, the second will be tried.

Testing this function may be a bit difficult, because it is not used yet.
I tested it by calling `get_webdav_door "$api"` just before the call to `api_call`.
To test the second method, I commented out the first `return` command in `get_webdav_door`.

With this function, we can start implementing uploads and downloads into Ada.

I'd also like to implement get_api_from_webdav. That would allow users to provide a WebDAV URL
if they don 't know the address of the API. But that is less urgent.

* Send debug output to stderr instead of stdout; print error when no WebDAV found
* The WebDAV door should be sent to stdout, that means that any other output should be sent to stderr to keep them separate.
* It's nice to print an error message when no WebDAV door could be found. This includes a hint to try with --debug, to speed up troubleshooting.

* get_webdav_door: add method: read WebDAV URL from Rclone config file is specified

* Be more polite please ;-)

* Return only when the method has succeeded.