With --whoami, show user quotas if set. With --whoami, ada will check whether the user has any quota (on their UID and on their main GID). If they have, the disk & tape quotas are shown, including percentages.
If the API returns 404 for both UID and GID quota, it means
quotas have not been set, so Ada does not show anything.

If the API returns other error messages, Ada will show them.

I'm thinking of implementing a `ada --quota` command which
shows all quotas, not only the user's own. That would be a next commit.