Add support for version locally stored
This PR adds support to read ADA version from a local file. This is
useful when the script is installed out of the context of a Git
repository and it is not possible to use the repository tag as version.

The script will read /etc/ada-version if it cannot determine a version
from a git repository.