Github action to update version number inside Ada script (#131)
* Add ADA_VERSION which can be automatically set by Github action

* Set version to 123-provide-reliable-ada---version

* Changed comment
Because I need to check if the version update works

* When pushing branch, add commit hash to the version.

* Reset version number to test Github Action

* Set version to 123-provide-reliable-ada---version

* Add commit hash to version also on branch push

* Remove commit hash from the version number
Adding the commit hash to the version number was a bad idea, because all the time the remote has a commit extra compared to your local branch. Then you get the error:
"Updates were rejected because the remote contains work that you do not have locally."
You can solve that by doing git push --force, but it's not very convenience (and probably not quite safe) to do that with every push.

* Pushing something to test what the Github action does

* Add comments to explain the ADA_VERSION values

* Clean up update-version-file.yml and ada-version
They are from a previous effort to automate version info

* Rename Update-ada-version.yml for consistency; improved sed command
* Renamed Update-ada-version.yml to update-ada-version.yml, because existing actions don't have capitals in their name either
* Changed sed command to be more compatible with MacOS:
  sed -i expression file ->
  sed expression -i file

* ${GITHUB_REF_NAME} already is either a tag or a branch name; if statement unnecessary.
Thanks Haili for reading the documentation :-)

* Make update_ada_version.yml work on MacOS and UNIX

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>
Co-authored-by: Haili Hu <hailihu@gmail.com>